+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### added
 - the combined media of all formsets is now available to the template context under the name `formset_media` 
+- the add button text can now be modified using the `add_text` keyword argument for `InlineFormsetRenderer`
 
 ## 0.1.0 (2023-08-31)